chore: update the dependencies in build.gradle
- androidx.appcompat
- com.google.android.material
- androidx.compose.material

The old dependencies did not contain the necessary library for
achieving the desired color theming. com.google.android.material
gives access to the following attributes
- colorPrimary
- colorPrimaryVariant
- colorOnPrimary which allows for the desired coloring effect.

The parent of the style element in themes.xml seems to draw from
appcompat which is why androidx.compat was included.

androidx.compose.material contains the Shapes class which may later be
used to change the shape of the button views. If not required, then
will be removed in future commits unless additional dependencies on
other compose.material classes are created.