feat: add behavior when met with clicks to specific buttons
Buttons:
- Numbers
	append number to the textview
- '('
	append the opening parenthesis to the textview and increase the
    global counter for opening parenthesis by one
- ')'
	check to see if the global counter for opening parenthesis is
    greater than or equal to one and if it is append it to the
    textview and decrease the global counter by one.
- '='
    if the last character is a valid ending character, then call a
    function that evaluates the entire textview to return back an
    answer. Otherwise, have the result textview display invalid and
    reset all global values.
- '.'
	if the correct conditions are met of the textview being empty or
	another decimal not having been placed before in a number,
	append the decimal to the textview.

In the prior iterations of the numbersAction function, only numerical
values were evaluated. The conditionals for evaluation for the
non-numerical characters were put aside for implementation in the
future. Since the calculator needs to evaluate based on infix notation
as well as be able to evaluate float values, these additional
conditionals were placed. However, there are some fixes that are
required regarding evaluation when there is a mismatch in the
number of opening to closing parenthesis. The program will crash when
asked to evaluate in this state.