feat: evaluate textview expression and return final answer
The following commit contains the code for evaluating the two stacks
(one for numbers and one for operation) to generate the answer for the
expression displayed in the working textview. Two helper functions
(quickMaths and finalStackEval) are used to solve
number/operator/number pairings (ex: 1+2). Another helper funtion
(opPrecedenceGreaterOrEqual) is used to determine the prescedence of
one operator vs another on the stack so that the infix notation is
evaluated in the correct manner to arrive at the desired answer.