fix: fix a crashing bug where due to misuse of variable
There was a bug that was causing the program to crash becuase it was
using the temporary variable for storing numbers to add blank numbers
to the stack. Since we store the numbers as strings before converting
to floats to add to the stack, calling toFloat() on an empty string
was crashing the program.